requirements.txt
importlib_metadata
pandas
pytest
python_dateutil
PyYAML
PyYAML
setuptools
Sphinx
svgwrite
dateutil
